Tarun, Rakesh | MOBN-1429 | Restructured notification test folder
